Refactor main menu and pause menu to streamline button layout; add game save and load functionality with JSON support
